requirements.txt
kafka-python==2.0.2
flask==3.0.2
flask-socketio==5.3.6
python-dotenv==1.0.1
gunicorn==21.2.0
eventlet==0.35.2 